pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "korgi"
version = "0.0.1"
description = "Cute AI TAs giving you lectures"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "anthropic>=0.40",
    "markitdown[all]>=0.0.1a",
    "typer>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
elevenlabs = ["elevenlabs>=1.0"]
moss = ["moss-tts-nano"]
dev = ["pytest>=8", "pytest-mock>=3"]

[project.scripts]
korgi = "korgi.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["korgi"]

[tool.pytest.ini_options]
testpaths = ["tests"]
